added Financial Indicators page UI
