+++ lib/presentation/pages/admin_profile/financial_indicators_page.dart
@@ -1,6 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:meta_app/core/utils/extensions/build_context_ext.dart';
+import 'package:meta_app/presentation/widgets/bot_demo_card.dart';
+import 'package:meta_app/presentation/widgets/level_card.dart';
 import 'package:meta_app/presentation/widgets/profile_header/header.dart';
+import 'package:meta_app/presentation/widgets/rights_reserved_footer.dart';
 
 class FinancialIndicatorsPage extends StatelessWidget {
   const FinancialIndicatorsPage({super.key});
@@ -10,9 +13,138 @@ class FinancialIndicatorsPage extends StatelessWidget {
     return Scaffold(
       appBar: ProfileHeader.admin(),
       backgroundColor: context.color.profilePageBackground,
-      body: Row(
+      body: SingleChildScrollView(
+        child: Container(
+          padding: const EdgeInsets.all(20),
+          decoration: BoxDecoration(
+            gradient: context.gradient.lightIndigo,
+          ),
+          child: Column(
+            children: const [
+              _BotDemoCards(),
+              SizedBox(height: 40),
+              _ReferralRewards(),
+              SizedBox(height: 80),
+              RightsReservedFooter(),
+            ],
+          ),
+        ),
+      ),
+    );
+  }
+}
+
+class _ReferralRewards extends StatelessWidget {
+  const _ReferralRewards({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      padding: const EdgeInsets.all(20),
+      decoration: BoxDecoration(
+        borderRadius: BorderRadius.circular(8),
+        color: context.color.profilePageBackground,
+      ),
+      child: Column(
         crossAxisAlignment: CrossAxisAlignment.start,
-        children: const [Text("admin")],
+        children: [
+          Text(
+            context.localizations.auroraUniverseReferralReward,
+            style: context.text.profileBotsDefault.copyWith(fontSize: 24),
+          ),
+          const SizedBox(height: 30),
+          const _LevelCards(),
+        ],
+      ),
+    );
+  }
+}
+
+class _BotDemoCards extends StatelessWidget {
+  const _BotDemoCards({Key? key}) : super(key: key);
+
+  List<String> _basicBenefits(BuildContext context) {
+    return [
+      context.localizations.twentyPositions,
+      context.localizations.portfolioManagement,
+      context.localizations.manualTrading,
+      context.localizations.allAvailableExchanges,
+    ];
+  }
+
+  List<String> _advancedBenefits(BuildContext context) {
+    return [
+      ..._basicBenefits(context),
+      context.localizations.allCoinsForSignals,
+      context.localizations.marketArbitrage,
+      context.localizations.algorithmIntelligence,
+    ];
+  }
+
+  List<Widget> _demoCards(BuildContext context) {
+    final localization = context.localizations;
+
+    return [
+      BotDemoCard(
+        title: localization.pioneer,
+        price: 20.39,
+        benefits: _basicBenefits(context),
+      ),
+      BotDemoCard(
+        title: localization.adventurer,
+        price: 40.22,
+        benefits: _basicBenefits(context),
+      ),
+      BotDemoCard(
+        title: localization.hero,
+        price: 78.25,
+        benefits: _advancedBenefits(context),
+      ),
+      BotDemoCard(
+        title: localization.pioneer,
+        price: 3000,
+        benefits: _advancedBenefits(context),
+      ),
+      BotDemoCard(
+        title: localization.adventurer,
+        price: 5000,
+        benefits: _advancedBenefits(context),
+      ),
+    ];
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return SizedBox(
+      height: 380,
+      width: double.infinity,
+      child: Align(
+        child: ListView.separated(
+          shrinkWrap: true,
+          scrollDirection: Axis.horizontal,
+          itemBuilder: (context, index) => _demoCards(context).elementAt(index),
+          itemCount: _demoCards(context).length,
+          separatorBuilder: (_, __) => const SizedBox(width: 20),
+        ),
+      ),
+    );
+  }
+}
+
+class _LevelCards extends StatelessWidget {
+  const _LevelCards({Key? key}) : super(key: key);
+
+  @override
+  Widget build(BuildContext context) {
+    return Wrap(
+      spacing: 20,
+      runSpacing: 16,
+      children: List.generate(
+        15,
+        (index) => LevelCard(
+          level: index + 1,
+          reward: index % 2 == 0 ? 300 : 500,
+        ),
       ),
     );
   }
