+++ lib/presentation/widgets/web_scroll_behaviour.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/gestures.dart';
+import 'package:flutter/material.dart';
+
+class WebScrollBehaviour extends MaterialScrollBehavior {
+  @override
+  Set<PointerDeviceKind> get dragDevices => {
+        PointerDeviceKind.touch,
+        PointerDeviceKind.mouse,
+      };
+}
